refactor: code optimizations
