+++ .pkgs/eslint-plugin-local/package.json
@@ -43,7 +43,7 @@
   "devDependencies": {
     "@local/configs": "workspace:*",
     "@types/react": "^19.1.2",
-    "@types/react-dom": "^19.1.2",
+    "@types/react-dom": "^19.1.3",
     "tsup": "^8.4.0"
   },
   "peerDependencies": {
