+++ apps/website/package.json
@@ -37,7 +37,7 @@
     "@eslint/markdown": "^6.4.0",
     "@local/configs": "workspace:*",
     "@mdx-js/mdx": "^3.1.0",
-    "@tailwindcss/postcss": "^4.1.4",
+    "@tailwindcss/postcss": "^4.1.5",
     "@theguild/remark-mermaid": "^0.3.0",
     "@tsconfig/next": "^2.0.3",
     "@tsconfig/node22": "^22.0.1",
@@ -46,7 +46,7 @@
     "@types/mdx": "^2.0.13",
     "@types/node": "^22.15.3",
     "@types/react": "^19.1.2",
-    "@types/react-dom": "^19.1.2",
+    "@types/react-dom": "^19.1.3",
     "autoprefixer": "^10.4.21",
     "dedent": "^1.5.3",
     "eslint": "^9.25.1",
@@ -58,7 +58,7 @@
     "eslint-plugin-unicorn": "^59.0.0",
     "importx": "^0.5.2",
     "postcss": "^8.5.3",
-    "tailwindcss": "^4.1.4",
+    "tailwindcss": "^4.1.5",
     "tailwindcss-animated": "^2.0.0",
     "typescript": "^5.8.3",
     "typescript-eslint": "^8.31.1"
