+++ examples/next-app/package.json
@@ -24,7 +24,7 @@
     "@tsconfig/strictest": "^2.0.5",
     "@types/node": "^22.15.3",
     "@types/react": "^19.1.2",
-    "@types/react-dom": "^19.1.2",
+    "@types/react-dom": "^19.1.3",
     "eslint": "^9.25.1",
     "eslint-config-flat-gitignore": "^2.1.0",
     "eslint-plugin-react-hooks": "^5.2.0",
