+++ examples/vite-react-dom-js-app/package.json
@@ -19,13 +19,13 @@
     "@eslint/config-inspector": "^1.0.2",
     "@eslint/js": "^9.25.1",
     "@types/react": "^19.1.2",
-    "@types/react-dom": "^19.1.2",
+    "@types/react-dom": "^19.1.3",
     "@vitejs/plugin-react": "^4.4.1",
     "eslint": "^9.25.1",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
     "globals": "^16.0.0",
-    "vite": "^6.3.3"
+    "vite": "^6.3.4"
   },
   "engines": {
     "node": ">=20.19.0"
