+++ examples/vite-react-dom-js-with-babel-eslint-parser-app/package.json
@@ -15,16 +15,16 @@
     "react-dom": "^19.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.26.10",
-    "@babel/eslint-parser": "^7.27.0",
-    "@babel/preset-env": "^7.26.9",
-    "@babel/preset-react": "^7.26.3",
+    "@babel/core": "^7.27.1",
+    "@babel/eslint-parser": "^7.27.1",
+    "@babel/preset-env": "^7.27.1",
+    "@babel/preset-react": "^7.27.1",
     "@eslint/config-inspector": "^1.0.2",
     "@eslint/js": "^9.25.1",
     "@types/babel__core": "~7.20.5",
     "@types/babel__preset-env": "~7.10.0",
     "@types/react": "^19.1.2",
-    "@types/react-dom": "^19.1.2",
+    "@types/react-dom": "^19.1.3",
     "@vitejs/plugin-react": "^4.4.1",
     "eslint": "^9.25.1",
     "eslint-plugin-react-debug": "workspace:*",
@@ -35,7 +35,7 @@
     "eslint-plugin-react-web-api": "workspace:*",
     "eslint-plugin-react-x": "workspace:*",
     "globals": "^16.0.0",
-    "vite": "^6.3.3"
+    "vite": "^6.3.4"
   },
   "engines": {
     "node": ">=20.19.0"
