+++ examples/vite-react-dom-with-ts-blank-eslint-parser-app/package.json
@@ -24,7 +24,7 @@
     "@tsconfig/strictest": "^2.0.5",
     "@tsconfig/vite-react": "^3.4.0",
     "@types/react": "^19.1.2",
-    "@types/react-dom": "^19.1.2",
+    "@types/react-dom": "^19.1.3",
     "@vitejs/plugin-react": "^4.4.1",
     "eslint": "^9.25.1",
     "eslint-plugin-react-hooks": "^5.2.0",
@@ -33,7 +33,7 @@
     "ts-blank-eslint-parser": "^0.4.3",
     "typescript": "^5.8.3",
     "typescript-eslint": "^8.31.1",
-    "vite": "^6.3.3"
+    "vite": "^6.3.4"
   },
   "engines": {
     "node": ">=20.19.0"
