+++ package.json
@@ -60,7 +60,7 @@
     "@tsconfig/strictest": "^2.0.5",
     "@types/node": "^22.15.3",
     "@types/react": "^19.1.2",
-    "@types/react-dom": "^19.1.2",
+    "@types/react-dom": "^19.1.3",
     "@typescript-eslint/parser": "^8.31.1",
     "@typescript-eslint/rule-tester": "^8.31.1",
     "@typescript-eslint/types": "^8.31.1",
@@ -109,7 +109,7 @@
     ],
     "overrides": {
       "@types/react": "^19.1.2",
-      "@types/react-dom": "^19.1.2",
+      "@types/react-dom": "^19.1.3",
       "cross-spawn": "^7.0.6",
       "esbuild": "^0.25.3",
       "lucide-react": "^0.503.0",
