+++ packages/plugins/eslint-plugin-react-debug/package.json
@@ -57,7 +57,7 @@
   "devDependencies": {
     "@local/configs": "workspace:*",
     "@types/react": "^19.1.2",
-    "@types/react-dom": "^19.1.2",
+    "@types/react-dom": "^19.1.3",
     "tsup": "^8.4.0"
   },
   "peerDependencies": {
