+++ packages/plugins/eslint-plugin-react-x/package.json
@@ -58,7 +58,7 @@
   "devDependencies": {
     "@local/configs": "workspace:*",
     "@types/react": "^19.1.2",
-    "@types/react-dom": "^19.1.2",
+    "@types/react-dom": "^19.1.3",
     "ts-api-utils": "^2.1.0",
     "tsup": "^8.4.0"
   },
