+++ packages/plugins/eslint-plugin-react-x/src/rules/no-unstable-default-props.ts
@@ -78,7 +78,7 @@ export function create(context: RuleContext<MessageID, []>): RuleListener {
           if (ER.isReactHookCall(construction.node)) {
             continue;
           }
-          const forbiddenType = AST.toDelimiterCaseType(right);
+          const forbiddenType = AST.toDelimiterFormat(right);
           context.report({
             messageId: "noUnstableDefaultProps",
             node: right,
