+++ packages/utilities/ast/src/__tests__/ast-hierarchy.test.ts
@@ -1,4 +1,4 @@
-import type { TSESTreeFunction } from "../ast-node";
+import type { TSESTreeFunction } from "../ast-node-types";
 
 import path from "node:path";
 import { parseForESLint } from "@typescript-eslint/parser";
@@ -9,7 +9,7 @@ import tsx from "dedent";
 import { describe, expect, it } from "vitest";
 import { getFixturesRootDir } from "../../../../../test";
 import { getNestedReturnStatements } from "../ast-hierarchy";
-import { isFunction } from "../ast-is";
+import { isFunction } from "../ast-node-is";
 
 function parse(code: string) {
   return parseForESLint(code, {
