+++ packages/utilities/ast/src/ast-expression.ts
@@ -1,8 +1,8 @@
 import type { TSESTree } from "@typescript-eslint/types";
-import type { TSESTreeTypeExpression } from "./ast-node";
+import type { TSESTreeTypeExpression } from "./ast-node-types";
 
 import { AST_NODE_TYPES as T } from "@typescript-eslint/types";
-import { isTypeExpression } from "./ast-is";
+import { isTypeExpression } from "./ast-node-is";
 
 /**
  * Recursively get the inner expression until it's not a TypeExpression
