+++ packages/utilities/ast/src/ast-line.ts
@@ -1,6 +1,6 @@
 import type { TSESTree } from "@typescript-eslint/types";
 import { AST_NODE_TYPES as T } from "@typescript-eslint/types";
-import { isOneOf } from "./ast-is";
+import { isOneOf } from "./ast-node-is";
 
 /**
  * Check if a node is multiline
