+++ packages/utilities/ast/src/ast-literal-is.ts
@@ -1,20 +1,21 @@
 import type { TSESTree } from "@typescript-eslint/types";
 import { AST_NODE_TYPES as T } from "@typescript-eslint/types";
 
-type LiteralType =
-  | "boolean"
-  | "null"
-  | "number"
-  | "regexp"
-  | "string";
-
 export function isLiteral(node: TSESTree.Node): node is TSESTree.Literal;
 export function isLiteral(node: TSESTree.Node, type: "boolean"): node is TSESTree.BooleanLiteral;
 export function isLiteral(node: TSESTree.Node, type: "null"): node is TSESTree.NullLiteral;
 export function isLiteral(node: TSESTree.Node, type: "number"): node is TSESTree.NumberLiteral;
 export function isLiteral(node: TSESTree.Node, type: "regexp"): node is TSESTree.RegExpLiteral;
 export function isLiteral(node: TSESTree.Node, type: "string"): node is TSESTree.StringLiteral;
-export function isLiteral(node: TSESTree.Node, type?: LiteralType) {
+export function isLiteral(
+  node: TSESTree.Node,
+  type?:
+    | "boolean"
+    | "null"
+    | "number"
+    | "regexp"
+    | "string",
+) {
   if (node.type !== T.Literal) return false;
   if (type == null) return true;
   switch (type) {
