+++ packages/utilities/ast/src/index.ts
@@ -1,18 +1,17 @@
 export * from "./ast-array-method";
 export * from "./ast-array-method-callback";
-export * from "./ast-array-pattern";
 export * from "./ast-class-id";
 export * from "./ast-expression";
 export * from "./ast-function-id";
 export * from "./ast-function-init-path";
 export * from "./ast-function-is";
 export * from "./ast-hierarchy";
-export * from "./ast-is";
 export * from "./ast-line";
 export * from "./ast-literal-is";
-export type * from "./ast-node";
 export * from "./ast-node-equal";
-export * from "./ast-process-env-node-env";
+export * from "./ast-node-format";
+export * from "./ast-node-is";
+export type * from "./ast-node-types";
+export * from "./ast-process-env";
 export * from "./ast-property-name";
 export * from "./ast-then";
-export * from "./ast-to-string";
